+++ packages/create-hydrogen/src/utils/versions.ts
@@ -1,0 +1,1 @@
+export const HYDROGEN_VERSION = '0.11.1-experimental.1'