+++ packages/create-hydrogen/templates/template-hydrogen-default/src/App.server.jsx
@@ -1,0 +1,27 @@
+import renderHydrogen from '@shopify/hydrogen/entry-server';
+import {Router, FileRoutes, ShopifyProvider} from '@shopify/hydrogen';
+import {Suspense} from 'react';
+import shopifyConfig from '../shopify.config';
+import DefaultSeo from './components/DefaultSeo.server';
+import NotFound from './components/NotFound.server';
+import LoadingFallback from './components/LoadingFallback';
+import CartProvider from './components/CartProvider.client';
+
+function App({routes, ...serverProps}) {
+  return (
+    <Suspense fallback={<LoadingFallback />}>
+      <ShopifyProvider shopifyConfig={shopifyConfig}>
+        <CartProvider>
+          <DefaultSeo />
+          <Router fallback={<NotFound />} serverProps={serverProps}>
+            <FileRoutes routes={routes} />
+          </Router>
+        </CartProvider>
+      </ShopifyProvider>
+    </Suspense>
+  );
+}
+
+const routes = import.meta.globEager('./routes/**/*.server.[jt](s|sx)');
+
+export default renderHydrogen(App, {shopifyConfig, routes});