+++ packages/create-hydrogen/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @shopify/create-hydrogen
 
+## 0.32.0
+
+### Minor Changes
+
+- c1705a5: Adds the following new flags for the create-hydrogen command
+
+  - template flag (-t): One of our templates such as, template-hydrogen-default, from the hydrogen /examples directory.
+  - name flag (-n): Dafaults to hydrogen-app
+  - path flag (-p)
+  - dependency-manager flag (-d): One of 'npm', 'yarn' or 'pnpm'
+  - shopify-cli-version flag (-s)
+  - hydrogen-version flag (-v)
+
+  Adds the following new Prompts for the create-hydrogen command
+
+  "Name your new Hydrogen storefront": (default: 'hydrogen-app')
+  "Choose a template": (default: 'template-hydrogen-minimal')
+
+  Uses new template functions from the @shopify/cli-kit package for scaffolding
+
+## 0.31.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @shopify/cli-kit@0.31.3
+  - @shopify/cli-hydrogen@0.31.3
+
+## 0.31.1
+
+### Patch Changes
+
+- Fix create-app and create-hydrogen that fail because "open" is missing
+
+## 0.31.0
+
+### Patch Changes
+
+- Updated dependencies [e701cf5]
+- Updated dependencies [563f225]
+  - @shopify/cli-kit@0.31.0
+  - @shopify/cli-hydrogen@0.31.0
+
 ## 0.30.2
 
 ### Patch Changes